requirements.txt
aiofiles
aiohttp
asyncio
beautifulsoup4
dnspython
ffmpeg-python
gitpython
hachoir
OpsAi
heroku3
httpx
humanize
lyricsgenius
motor
numpy
pillow
psutil
py-tgcalls==0.9.2
pykeyboard
pyrogram==2.0.106
python-dotenv
pyyaml
requests
speedtest-cli
spotipy
telegraph
tgcrypto
unidecode
yt-dlp
youtube-search
youtube-search-python
